+++ lib/xdot/xdot.h
@@ -53,7 +53,7 @@ typedef struct {
 typedef struct {
   double x0, y0;
   double x1, y1;
-  int n_stops;
+  unsigned n_stops;
   xdot_color_stop *stops;
 } xdot_linear_grad;